% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_practice.R
\name{bp}
\alias{bp}
\alias{bp_extract_signatures}
\title{A Best Practice for Signature Extraction and Exposure (Activity) Attribution}
\usage{
bp_extract_signatures(
  nmf_matrix,
  range = 2:5,
  n_bootstrap = 20L,
  n_nmf_run = 50,
  RTOL = 0.001,
  min_contribution = 0,
  cores = min(4L, future::availableCores()),
  seed = 123456L,
  handle_hyper_mutation = TRUE,
  report_integer_exposure = TRUE,
  mpi_platform = FALSE
)
}
\arguments{
\item{nmf_matrix}{a \code{matrix} used for NMF decomposition with rows indicate samples and columns indicate components.}

\item{range}{a \code{numeric} vector containing the ranks of factorization to try. Note that duplicates are removed
and values are sorted in increasing order. The results are notably returned in this order.}

\item{n_bootstrap}{number of bootstrapped (resampling) catalogs used.
When it is \code{0}, the original (input) mutation catalog is used for NMF decomposition.}

\item{n_nmf_run}{number of NMF runs for each bootstrapped or original catalog.
At default, in total n_bootstrap x n_nmf_run (i.e. 1000) NMF runs are used
for the task.}

\item{RTOL}{a threshold proposed by Nature Cancer paper to control how to
filter solutions of NMF. Default is \verb{0.1\%}, only NMF solutions with KLD
(KL deviance) <= \verb{100.1\%} minimal KLD are kept.}

\item{min_contribution}{a component contribution threshold to filer out small
contributed components.}

\item{cores}{number of cpu cores to run NMF.}

\item{seed}{a random seed to make reproducible result.}

\item{handle_hyper_mutation}{default is \code{TRUE}, handle hyper-mutant samples.}

\item{report_integer_exposure}{default is \code{TRUE}, report integer signature
exposure by bootstrapping technique.}

\item{mpi_platform}{keep for implementation.}
}
\value{
It depends on the called function.
}
\description{
\itemize{
\item \code{bp_extract_signatures()} for extracting signatures.
}
}
