% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.hg19}
\alias{centromeres.hg19}
\title{Location of Centromeres at Genome Build hg19}
\format{
A data.frame
}
\source{
Generate from UCSC gold path
}
\description{
Location of Centromeres at Genome Build hg19
}
\examples{
data(centromeres.hg19)
}
