% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.hg38}
\alias{centromeres.hg38}
\title{Location of Centromeres at Genome Build hg38}
\format{
A data.frame
}
\source{
Generate from Genome Reference Consortium
}
\description{
Location of Centromeres at Genome Build hg38
}
\examples{
data(centromeres.hg38)
}
