% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.mm10}
\alias{centromeres.mm10}
\title{Location of Centromeres at Genome Build mm10}
\format{
A data.frame
}
\source{
Generate from \url{https://hgdownload.soe.ucsc.edu/goldenPath/mm10/database/gap.txt.gz}
}
\description{
Location of Centromeres at Genome Build mm10
}
\examples{
data(centromeres.mm10)
}
