% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.mm9}
\alias{centromeres.mm9}
\title{Location of Centromeres at Genome Build mm9}
\format{
A data.frame
}
\source{
Generate from \url{https://hgdownload.soe.ucsc.edu/goldenPath/mm9/database}
}
\description{
Location of Centromeres at Genome Build mm9
}
\examples{
data(centromeres.mm9)
}
