% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.T2T}
\alias{centromeres.T2T}
\title{Location of Centromeres at Genome Build T2T}
\format{
A data.frame
}
\source{
from T2T study
}
\description{
Location of Centromeres at Genome Build T2T
}
\examples{
data(centromeres.T2T)
}
