% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize.hg19}
\alias{chromsize.hg19}
\title{Chromosome Size of Genome Build hg19}
\format{
A data.frame
}
\source{
Generate from UCSC gold path
}
\description{
Chromosome Size of Genome Build hg19
}
\examples{
data(chromsize.hg19)
}
