% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize.hg38}
\alias{chromsize.hg38}
\title{Chromosome Size of Genome Build hg38}
\format{
A data.frame
}
\source{
Generate from UCSC gold path
}
\description{
Chromosome Size of Genome Build hg38
}
\examples{
data(chromsize.hg38)
}
