% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize.mm10}
\alias{chromsize.mm10}
\title{Chromosome Size of Genome Build mm10}
\format{
A data.frame
}
\source{
Generate from UCSC gold path \url{http://hgdownload.cse.ucsc.edu/goldenPath/mm10/bigZips/mm10.chrom.sizes}
}
\description{
Chromosome Size of Genome Build mm10
}
\examples{
data(chromsize.mm10)
}
