% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize.mm9}
\alias{chromsize.mm9}
\title{Chromosome Size of Genome Build mm9}
\format{
A data.frame
}
\source{
Generate from UCSC gold path \url{http://hgdownload.cse.ucsc.edu/goldenPath/mm9/bigZips/mm9.chrom.sizes}
}
\description{
Chromosome Size of Genome Build mm9
}
\examples{
data(chromsize.mm9)
}
