% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize.T2T}
\alias{chromsize.T2T}
\title{Chromosome Size of Genome Build T2T}
\format{
A data.frame
}
\source{
from T2T study
}
\description{
Chromosome Size of Genome Build T2T
}
\examples{
data(chromsize.T2T)
}
