% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CN.features}
\alias{CN.features}
\title{Classification Table of Copy Number Features Devised by Wang et al. for Method 'W'}
\format{
A \code{data.table} with "sigminer.features" class name
}
\source{
Generate from code under data_raw/
}
\description{
Classification Table of Copy Number Features Devised by Wang et al. for Method 'W'
}
\examples{
data(CN.features)
}
