% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf_series.R
\name{cnv_autoCaptureSignatures}
\alias{cnv_autoCaptureSignatures}
\title{Auto-capture signature and coresponding exposure}
\usage{
cnv_autoCaptureSignatures(sample_by_component, nTry = 12, nrun = 10,
  cores = 1, seed = 123456, plot = TRUE,
  consensusmap_name = "nmf_consensus", testRandom = TRUE,
  nmfalg = "brunet")
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \link{cnv_generateSbCMatrix} function.}

\item{nTry}{the maximal tried number of signatures, default is 12.
Of note, this value should far less than number of features or samples.}

\item{nrun}{the number of run to perform for each value in range of 2 to \code{nTry}, default is 10.
According to \strong{NMF} package documentation, \code{nrun} set to 30~50 is enough to achieve robust result.}

\item{cores}{number of compute cores to run this task.
You can use \code{\link[parallel:detectCores]{parallel::detectCores()}} function to check how
many cores you can use.}

\item{seed}{seed number.}

\item{plot}{logical. If \code{TRUE}, plot rank survey.}

\item{consensusmap_name}{a character, basename of consensus map output path.}

\item{testRandom}{Should generate random data from input to test measurements. Default is \code{TRUE}.}

\item{nmfalg}{specification of the NMF algorithm.}
}
\value{
a \code{list} contains results of NMF best rank survey, run, signature matrix, exposure list etc..
}
\description{
this is a wrapper of \code{cnv_chooseSigNumber}, \code{cnv_extractSignatures}
and \code{cnv_quantifySigExposure} these three functions.
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## generate copy-number features
tcga_features = get_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = get_components(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = cnv_generateSbCMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey13
tcga_results = cnv_autoCaptureSignatures(tcga_sample_component_matrix, nrun=10, cores = 1)
}
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_chooseSigNumber}},
  \code{\link{cnv_extractSignatures}},
  \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}}
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
\concept{CNV analysis functions}
