% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{cnv_plotDistributionProfile}
\alias{cnv_plotDistributionProfile}
\title{Plot copy number distribution either by length or chromosome}
\usage{
cnv_plotDistributionProfile(data, rm_normal = TRUE, mode = c("ld",
  "cd"), fill = FALSE, scale_chr = TRUE, genome_build = c("hg19",
  "hg38"))
}
\arguments{
\item{data}{a \code{data.frame}, result of \code{get_LengthFraction} function.}

\item{rm_normal}{logical. Whether removel normal copy (i.e. "segVal" equals 2), default is \code{TRUE}.}

\item{mode}{either "ld" for distribution by CN length or "cd" for distribution by chromosome.}

\item{fill}{when \code{mode} is "cd", if \code{fill} is \code{TRUE}, plot percentage instead of count.}

\item{scale_chr}{logical. If \code{TRUE}, normalize count to per Megabase unit.}
}
\value{
a ggplot object
}
\description{
Plot copy number distribution either by length or chromosome
}
\examples{
\dontrun{
cnv_plotLengthSummary(data, mode = "cd")
}
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}},
  \code{\link{get_LengthFraction}}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
\concept{CNV analysis functions}
