% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{cnv_plotFeatureDistribution}
\alias{cnv_plotFeatureDistribution}
\title{Plot copy number feature distribution}
\usage{
cnv_plotFeatureDistribution(features, ylab = "", ...)
}
\arguments{
\item{features}{a \code{list} generate from \code{cnv_derivefeatures} function.}

\item{ylab}{lab of y axis.}

\item{...}{other options pass to \code{\link[cowplot]{plot_grid}} function of \code{cowplot} package.}
}
\value{
a ggplot object
}
\description{
Plot copy number feature distribution
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}},
  \code{\link{get_LengthFraction}}
}
\concept{CNV analysis functions}
