% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{cnv_plotSignatures}
\alias{cnv_plotSignatures}
\title{Plot CNV signatures}
\usage{
cnv_plotSignatures(Res = NULL, contributions = FALSE, color = NULL,
  patient_order = NULL, font_size = 1.2, show_title = TRUE,
  axis_lwd = 2, title_size = 1.3, show_barcodes = FALSE,
  barcode_size = 0.5, yaxisLim = 0.3, ...)
}
\arguments{
\item{Res}{a result \code{list} generated from CNV signature pipeline or a \code{NMFfitX} object from \code{NMF} package.}

\item{contributions}{if \code{TRUE}, plot contributions instead of signatures.}

\item{color}{user custom colors.}

\item{patient_order}{user custom patient order.}

\item{font_size}{font size.}

\item{show_title}{if \code{TRUE}, show title.}

\item{axis_lwd}{line width of axis.}

\item{title_size}{title size.}

\item{show_barcodes}{if \code{TRUE}, show barcode.}

\item{barcode_size}{barcode size when plot contribution.}

\item{yaxisLim}{limit of y axis.}

\item{...}{other options pass to \code{barplot}.}
}
\value{
Nothing
}
\description{
Plot CNV signature from multiple result objects from \code{VSHunter} package.
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_autoCaptureSignatures}},
  \code{\link{cnv_chooseSigNumber}},
  \code{\link{cnv_extractSignatures}},
  \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}}
}
\concept{CNV analysis functions}
