% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{CopyNumber-class}
\alias{CopyNumber-class}
\alias{CopyNumber}
\title{Class CopyNumber}
\description{
S4 class for storing summarized absolute copy number profile.
}
\section{Slots}{

\describe{
\item{\code{data}}{data.table of absolute copy number calling.}

\item{\code{genome_build}}{genome build version, should be one of 'hg19' or 'hg38'.}

\item{\code{annotation}}{data.table of annotation for copy number segments.}

\item{\code{dropoff.segs}}{data.table of copy number segments dropped from raw input.}

\item{\code{clinical.data}}{data associated with each sample in copy number profile.}
}}

