% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{cosine}
\alias{cosine}
\title{Calculate Cosine Measures}
\usage{
cosine(x, y)
}
\arguments{
\item{x}{a numeric vector or matrix with column representing vector to calculate similarity.}

\item{y}{must be same format as \code{x}.}
}
\value{
a numeric value or \code{matrix}.
}
\description{
Calculate Cosine Measures
}
\examples{
x <- c(1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0)
y <- c(0, 0, 1, 1, 1, 1, 1, 0, 1, 0, 0, 0)
z1 <- cosine(x, y)
z1
z2 <- cosine(matrix(x), matrix(y))
z2
}
