% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands.hg19}
\alias{cytobands.hg19}
\title{Location of Chromosome Cytobands at Genome Build hg19}
\format{
A data.frame
}
\source{
from UCSC
}
\description{
Location of Chromosome Cytobands at Genome Build hg19
}
\examples{
data(cytobands.hg19)
}
