% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands.hg38}
\alias{cytobands.hg38}
\title{Location of Chromosome Cytobands at Genome Build hg38}
\format{
A data.frame
}
\source{
from UCSC
}
\description{
Location of Chromosome Cytobands at Genome Build hg38
}
\examples{
data(cytobands.hg38)
}
