% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands.mm10}
\alias{cytobands.mm10}
\title{Location of Chromosome Cytobands at Genome Build mm10}
\format{
A data.frame
}
\source{
from UCSC \url{http://hgdownload.cse.ucsc.edu/goldenpath/mm10/database/cytoBand.txt.gz}
}
\description{
Location of Chromosome Cytobands at Genome Build mm10
}
\examples{
data(cytobands.mm10)
}
