% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands.mm9}
\alias{cytobands.mm9}
\title{Location of Chromosome Cytobands at Genome Build mm9}
\format{
A data.frame
}
\source{
from UCSC \url{http://hgdownload.cse.ucsc.edu/goldenpath/mm9/database/cytoBand.txt.gz}
}
\description{
Location of Chromosome Cytobands at Genome Build mm9
}
\examples{
data(cytobands.mm9)
}
