% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytobands.T2T}
\alias{cytobands.T2T}
\title{Location of Chromosome Cytobands at Genome Build T2T}
\format{
A data.frame
}
\source{
from T2T study
}
\description{
Location of Chromosome Cytobands at Genome Build T2T
}
\examples{
data(cytobands.T2T)
}
