% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive}
\alias{derive}
\alias{derive.CopyNumber}
\alias{derive.MAF}
\title{Derive Variation Matrix from Mutation Profile}
\usage{
derive(object, ...)

\method{derive}{CopyNumber}(object, reference_components = FALSE,
  cores = 1, seed = 123456, min_comp = 2, max_comp = 10,
  min_prior = 0.001, model_selection = "BIC", threshold = 0.1,
  nrep = 1, niter = 1000, rowIter = 1000, keep_only_matrix = FALSE,
  ...)

\method{derive}{MAF}(object, ref_genome = NULL, prefix = NULL,
  add = TRUE, ignoreChr = NULL, useSyn = TRUE,
  keep_only_matrix = FALSE, ...)
}
\arguments{
\item{object}{a \link{CopyNumber} object or \link{MAF} object.}

\item{...}{custom setting for operating object. Detail see S3 method for
corresponding class (e.g. \code{CopyNumber}).}

\item{reference_components}{default is \code{FALSE}, calculate mixture components
from \link{CopyNumber} object.}

\item{cores}{number of compute cores to run this task.
You can use \code{\link[parallel:detectCores]{parallel::detectCores()}} function to check how
many cores you can use.}

\item{seed}{seed number.}

\item{min_comp}{minimal number of components to fit, default is 2.
Can also be a vector with length 6, which apply to each feature.}

\item{max_comp}{maximal number of components to fit, default is 10.
Can also be a vector with length 6, which apply to each feature.}

\item{min_prior}{the minimum relative size of components, default is 0.001.
Details about custom setting please refer to \strong{flexmix} package.}

\item{model_selection}{model selection strategy, default is 'BIC'.
Details about custom setting please refer to \strong{flexmix} package.}

\item{threshold}{default is \code{0.1}. Sometimes, the result components
include adjacent distributions with similar mu
(two and more distribution are very close), we use this threshold
to obtain a more meaningful fit with less components.}

\item{nrep}{number of run times for each value of component,
keep only the solution with maximum likelihood.}

\item{niter}{the maximum number of iterations.}

\item{rowIter}{step size of iteration for rows of ech CNV feature.}

\item{keep_only_matrix}{if \code{TRUE}, keep only matrix for signature extraction.}

\item{ref_genome}{BSgenome object or name of the installed BSgenome package. Example: BSgenome.Hsapiens.UCSC.hg19
Default NULL, tries to auto-detect from installed genomes.}

\item{prefix}{Prefix to add or remove from contig names in MAF file.}

\item{add}{If prefix is used, default is to add prefix to contig names in MAF file. If false prefix will be removed from contig names.}

\item{ignoreChr}{Chromsomes to ignore from analysis. e.g. chrM}

\item{useSyn}{Logical. Whether to include synonymous variants in analysis. Defaults to TRUE}
}
\value{
a \code{list} contains a \code{matrix} used for NMF de-composition.
}
\description{
Generate a matrix for NMF de-composition and more. This is a generic function,
so it can be further extended to other mutation cases.
}
\section{Methods (by class)}{
\itemize{
\item \code{CopyNumber}: Derive copy number features, components and component-by-sample matrix

\item \code{MAF}: Derive SBS mutation features, components and component-by-sample matrix
}}

\examples{
\donttest{
# Load copy number object
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))
# Prepare copy number signature analysis
cn_prepare <- derive(cn)
}
\donttest{
# Prepare mutational signature analysis
laml.maf <- system.file("extdata", "tcga_laml.maf.gz", package = "maftools")
laml <- read_maf(maf = laml.maf)
library(BSgenome.Hsapiens.UCSC.hg19)
mt_prepare <- derive(
  laml,
  ref_genome = "BSgenome.Hsapiens.UCSC.hg19",
  prefix = "chr", add = TRUE, useSyn = TRUE
)
}
}
\references{
Macintyre, Geoff, et al. "Copy number signatures and mutational
processes in ovarian carcinoma." Nature genetics 50.9 (2018): 1262.

Mayakonda, Anand, et al. "Maftools: efficient and comprehensive analysis of somatic variants in cancer." Genome research 28.11 (2018): 1747-1756.
}
\author{
Shixiang Wang
}
