% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_correlation.R
\name{draw_features_corrplot}
\alias{draw_features_corrplot}
\title{Draw corrplot for signature activity and other features}
\usage{
draw_features_corrplot(tidy_cor, features_list, breaks_count = c(0L,
  200L, 400L, 600L, 800L, 1020L), breaks_score = c(-Inf, -16, -10, -5,
  -1.3, 1.3, 5, 10, 16, Inf))
}
\arguments{
\item{tidy_cor}{data returned by \code{\link[=sig_tidy_correlation]{sig_tidy_correlation()}}.}

\item{features_list}{a character vector contains features want to be plotted.}

\item{breaks_count}{breaks for sample count.}

\item{breaks_score}{breaks for score.}
}
\value{
a \code{ggplot2} object
}
\description{
Draw corrplot for signature activity and other features
}
