% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{draw_sig_enrichment}
\alias{draw_sig_enrichment}
\title{Plot signature enrichment result}
\usage{
draw_sig_enrichment(enrich_res, pVal = 0.05, cols = NULL,
  addGeneDist = 0, annoFontSize = 0.8, geneFontSize = 0.8,
  legendFontSize = 0.8, title = NULL, showTitle = TRUE,
  showLegend = TRUE)
}
\arguments{
\item{enrich_res}{results from \link{sig_group_enrichment},
or \link[maftools:clinicalEnrichment]{maftools::clinicalEnrichment} or \link[maftools:signatureEnrichment]{maftools::signatureEnrichment}.}

\item{pVal}{Default 0.05}

\item{cols}{named vector of colors for factor in a clinical feature. Default NULL}

\item{addGeneDist}{Distance added to Gene label. Default is 0.}

\item{annoFontSize}{cex for annotation font size. Default 0.8}

\item{geneFontSize}{cex for gene font size. Default 0.8}

\item{legendFontSize}{cex for legend font size. Default 0.8}

\item{title}{title for plot.}

\item{showTitle}{Default TRUE}

\item{showLegend}{if \code{TRUE}, show legend.}
}
\value{
a base plot
}
\description{
A modified version of \link[maftools:plotEnrichmentResults]{maftools::plotEnrichmentResults}.
}
