% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_component_strand_bias.R
\name{enrich_component_strand_bias}
\alias{enrich_component_strand_bias}
\title{Performs Strand Bias Enrichment Analysis for a Given Sample-by-Component Matrix}
\usage{
enrich_component_strand_bias(mat)
}
\arguments{
\item{mat}{a sample-by-component matrix from \link{sig_tally} with strand bias labels "T:" and "B:".}
}
\value{
a \code{data.table} sorted by \code{p_value}.
}
\description{
See \link{sig_tally} for examples.
}
