% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{filter_enrichment}
\alias{filter_enrichment}
\title{Filter enrichment result}
\usage{
filter_enrichment(enrich, subset = TRUE, type = c("groupwise",
  "pairwise"))
}
\arguments{
\item{enrich}{results from \link{sig_group_enrichment},
or \link[maftools:clinicalEnrichment]{maftools::clinicalEnrichment} or \link[maftools:signatureEnrichment]{maftools::signatureEnrichment}.}

\item{subset}{logical expression used to subset data according to \code{type}.}

\item{type}{data type to subset, either 'groupwise' or 'pairwise'.}
}
\value{
a filtered enrichment result
}
\description{
Filter enrichment result
}
