% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{GenomicVariation-class}
\alias{GenomicVariation-class}
\alias{GenomicVariation}
\title{Class GenomicVariation}
\description{
S4 class for storing summarized genomic variation profile.
}
\section{Slots}{

\describe{
\item{\code{CopyNumber}}{a \code{CopyNumber} object.}

\item{\code{MAF}}{a \code{MAF} object.}

\item{\code{clinical.data}}{data associated with each sample in copy number profile
and MAF. This slot is intersection of \code{clinical.data} slot in both \code{CopyNumber}
object and \code{MAF} object.}
}}

