% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_ArmLocation}
\alias{get_ArmLocation}
\title{Get chromosome arm location}
\usage{
get_ArmLocation(genome_build = c("hg19", "hg38"))
}
\arguments{
\item{genome_build}{genome build version, should be 'hg19' or 'hg38'.}
}
\description{
Get chromosome arm location
}
\seealso{
Other internal calculation function series: \code{\link{get_LengthFraction}},
  \code{\link{get_cnlist}}, \code{\link{get_components}},
  \code{\link{get_context}}, \code{\link{get_features}},
  \code{\link{get_matrix}}
}
\concept{internal calculation function series}
