% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cn_ploidy.R
\name{get_cn_ploidy}
\alias{get_cn_ploidy}
\title{Get Ploidy from Absolute Copy Number Profile}
\usage{
get_cn_ploidy(data)
}
\arguments{
\item{data}{a \link{CopyNumber} object or a \code{data.frame} containing at least 'chromosome', 'start',
'end', 'segVal' these columns.}
}
\value{
a value or a \code{data.table}
}
\description{
Get Ploidy from Absolute Copy Number Profile
}
\examples{
# Load copy number object
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))

df <- get_cn_ploidy(cn)
df
}
