% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_cnlist}
\alias{get_cnlist}
\title{Extract copy number profile as list from CopyNumber object}
\usage{
get_cnlist(CopyNumber)
}
\arguments{
\item{CopyNumber}{a \link{CopyNumber} object.}
}
\value{
a \code{list}
}
\description{
Extract copy number profile as list from CopyNumber object
}
\seealso{
Other internal calculation function series: \code{\link{get_ArmLocation}},
  \code{\link{get_LengthFraction}},
  \code{\link{get_components}}, \code{\link{get_context}},
  \code{\link{get_features}}, \code{\link{get_matrix}}
}
\author{
Shixiang Wang
}
\concept{internal calculation function series}
