% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_annotation.R
\name{get_genome_annotation}
\alias{get_genome_annotation}
\title{Get Genome Annotation}
\usage{
get_genome_annotation(data_type = c("chr_size", "centro_loc"),
  chrs = paste0("chr", c(1:22, "X", "Y")), genome_build = c("hg19",
  "hg38"), rm_extra = TRUE)
}
\arguments{
\item{data_type}{'chr_size' for chromosome size,
'centro_loc' for location of centromeres}

\item{chrs}{chromosomes start with 'chr'}

\item{genome_build}{one of 'hg19', 'hg38'}

\item{rm_extra}{if \code{TRUE}, remove extra contigs}
}
\value{
a \code{data.frame} containing annotation data
}
\description{
Get Genome Annotation
}
