% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_signature.R
\name{get_group_signature}
\alias{get_group_signature}
\title{Find Enriched Signature in Each Sample Group}
\usage{
get_group_signature(group_df, Signature)
}
\arguments{
\item{group_df}{a \code{data.frame} containing columns 'sample' and 'group',
it can be obtained from \code{\link[=get_groups]{get_groups()}}.}

\item{Signature}{a \code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract},
or just a raw exposure matrix with column representing samples (patients) and row
representing signatures.}
}
\value{
a \code{data.frame}
}
\description{
Find Enriched Signature in Each Sample Group
}
\examples{
\donttest{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
# Extract copy number signatures
library(NMF)
sig <- sig_extract(cn_prepare$nmf_matrix, 2, nrun = 10)

group_df = get_groups(sig, match_consensus = TRUE)
get_group_signature(group_df, sig)
}
}
