% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_groups.R
\name{get_groups}
\alias{get_groups}
\title{Get Sample Groups from Signature Decomposition Information}
\usage{
get_groups(Signature, method = c("exposure", "consensus", "samples"),
  match_consensus = FALSE)
}
\arguments{
\item{Signature}{a \code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract}.}

\item{method}{grouping method, could be one of the following:
\itemize{
\item 'exposure' - the default method, which assigns a sample into a group whose signature exposure
is dominant
\item 'consensus' - returns the cluster membership based on the hierarchical clustering of the consensus matrix,
it can only be used for the result obtained by \code{\link[=sig_extract]{sig_extract()}} with multiple runs using \strong{NMF} package.
\item 'samples' - returns the cluster membership based on the contribution of signature to each sample
using \strong{NMF} package.
}}

\item{match_consensus}{only used when the \code{method} is 'consensus'.
If \code{TRUE}, the result will match order as shown in consensus map.}
}
\value{
a \code{data.table} object
}
\description{
One of key results from signature analysis is to cluster samples into different
groups. This function takes \code{Signature} object as input
and return the membership in each cluster.
}
\examples{
\donttest{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
# Extract copy number signatures
library(NMF)
sig <- sig_extract(cn_prepare$nmf_matrix, 2, nrun = 10)

# Methods 'consensus' and 'samples' are from NMF::predict()
get_groups(sig)
get_groups(sig, method = "consensus", match_consensus = TRUE)
get_groups(sig, method = "samples")
}
}
\seealso{
\code{\link[NMF:predict]{NMF::predict()}}
}
