% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intersect_size.R
\name{get_intersect_size}
\alias{get_intersect_size}
\title{Get Overlap Size between Interval x and y}
\usage{
get_intersect_size(x.start, x.end, y.start, y.end)
}
\arguments{
\item{x.start}{start position of interval x.}

\item{x.end}{start position of interval x.}

\item{y.start}{start position of interval x.}

\item{y.end}{start position of interval x.}
}
\value{
a numeric vector.
}
\description{
Get Overlap Size between Interval x and y
}
\examples{
o1 <- get_intersect_size(1, 5, 3, 20)
o1
o2 <- get_intersect_size(3, 20, 1, 10)
o2
o3 <- get_intersect_size(c(1, 2, 1), c(10, 4, 6), c(4, 2, 5), c(10, 3, 22))
o3

}
