% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_LengthFraction}
\alias{get_LengthFraction}
\title{Calculate length fraction profile of copy number}
\usage{
get_LengthFraction(CN_data, genome_build = c("hg19", "hg38"),
  seg_cols = c("Chromosome", "Start.bp", "End.bp", "modal_cn"),
  samp_col = "sample")
}
\arguments{
\item{genome_build}{genome build version, should be one of 'hg19' or 'hg38'.}

\item{seg_cols}{four characters used to specify chromosome, start position,
end position and copy number value in \code{input}, respectively.
Default use names from ABSOLUTE calling result.}

\item{samp_col}{a character used to specify the sample column name. If \code{input}
is a directory and cannot find \code{samp_col}, sample names will use file names
(set this parameter to \code{NULL} is recommended in this case).}
}
\value{
a data frame
}
\description{
Calculate length fraction profile of copy number
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
\concept{CNV analysis functions}
