% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_exposure.R
\name{get_sig_exposure}
\alias{get_sig_exposure}
\title{Get Signature Exposure from 'Signature' Object}
\usage{
get_sig_exposure(Signature, type = c("absolute", "relative"))
}
\arguments{
\item{Signature}{a \code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract},
or just a raw exposure matrix with column representing samples (patients) and row
representing signatures.}

\item{type}{'absolute' for signature exposure and 'relative' for signature relative exposure.}
}
\value{
a \code{tibble}
}
\description{
Get Signature Exposure from 'Signature' Object
}
\examples{
# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Get signature exposure
get_sig_exposure(sig2)
get_sig_exposure(sig2, type = 'relative')
}
