% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_feature_association.R
\name{get_sig_feature_association}
\alias{get_sig_feature_association}
\title{Calculate Association between Signature Exposures and Other Features}
\usage{
get_sig_feature_association(data, cols_to_sigs, cols_to_summary,
  type = "ca", method_co = c("spearman", "pearson", "kendall"),
  method_ca = stats::wilcox.test, min_n = 0.01, verbose = FALSE, ...)
}
\arguments{
\item{data}{a \code{data.frame} contains signature exposures and other features}

\item{cols_to_sigs}{colnames for signature exposure}

\item{cols_to_summary}{colnames for other features}

\item{type}{a character vector contains 'ca' for categorical variable and 'co' for continuous variable,
it must have same length as \code{cols_to_summary}.}

\item{method_co}{method for continuous variable, default is "spearman", could also be "pearson" and "kendall".}

\item{method_ca}{method for categorical variable, default is "wilcox.test"}

\item{min_n}{a minimal fraction (e.g. 0.01) or a integer number (e.g. 10) for filtering some variables with few positive events.
Default is 0.01.}

\item{verbose}{if \code{TRUE}, print extra message.}

\item{...}{other arguments passing to test functions, like \code{cor.test}.}
}
\value{
a \code{list}
}
\description{
See \link{get_tidy_association} for cleaning association result.
}
\seealso{
\link{get_tidy_association}
}
