% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reconstructed_similarity.R
\name{get_sig_rec_similarity}
\alias{get_sig_rec_similarity}
\title{Get Reconstructed Profile Cosine Similarity, RSS, etc.}
\usage{
get_sig_rec_similarity(Signature, nmf_matrix)
}
\arguments{
\item{Signature}{a \code{Signature} object.}

\item{nmf_matrix}{a \code{matrix} used for NMF decomposition with rows indicate samples and columns indicate components.}
}
\value{
a \code{data.table}.
}
\description{
See \link{bp_extract_signatures} for examples.
}
