% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_association.R
\name{get_tidy_association}
\alias{get_tidy_association}
\title{Get Tidy Signature Association Results}
\usage{
get_tidy_association(cor_res)
}
\arguments{
\item{cor_res}{data returned by \code{\link[=get_sig_feature_association]{get_sig_feature_association()}}}
}
\value{
a \code{data.frame}
}
\description{
Get Tidy Signature Association Results
}
\seealso{
\link{get_sig_feature_association}
}
