% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tidy_parameter.R
\name{get_tidy_parameter}
\alias{get_tidy_parameter}
\title{Get Tidy Parameter from Flexmix Model}
\usage{
get_tidy_parameter(x)
}
\arguments{
\item{x}{a \code{flexmix} object or a \code{list} of \code{flexmix} objects.}
}
\value{
a \code{tibble}.
}
\description{
When users derive copy number features, it is useful to know the parameters
of the fit components, including mean, sd and coefficient of variation.
This function is used by \link{sig_feed} function and
exported to users for extra usage.
}
\examples{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
# Get all parameters
get_tidy_parameter(cn_prepare$components)
# Get parameters for segsize feature
get_tidy_parameter(cn_prepare$components$segsize)
}
