% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_hyper_mutation.R
\name{handle_hyper_mutation}
\alias{handle_hyper_mutation}
\title{Handle Hypermutant Samples}
\usage{
handle_hyper_mutation(nmf_matrix)
}
\arguments{
\item{nmf_matrix}{a \code{matrix} used for NMF decomposition with rows indicate samples and columns indicate components.}
}
\value{
a \code{matrix}.
}
\description{
This can be used for SNV/INDEL count matrix. For copy number analysis,
please skip it.
}
\references{
Kim, Jaegil, et al. "Somatic ERCC2 mutations are associated with a distinct genomic signature in urothelial tumors."
Nature genetics 48.6 (2016): 600.
}
