% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_bootstrap}
\alias{output_bootstrap}
\title{Output Signature Bootstrap Fitting Results}
\usage{
output_bootstrap(x, result_dir, mut_type = "SBS")
}
\arguments{
\item{x}{result from \link{sig_fit_bootstrap_batch}.}

\item{result_dir}{a result directory.}

\item{mut_type}{one of 'SBS', 'DBS', 'ID' or 'CN'.}
}
\value{
Nothing.
}
\description{
Output Signature Bootstrap Fitting Results
}
