% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_fit}
\alias{output_fit}
\title{Output Signature Fitting Results}
\usage{
output_fit(x, result_dir, mut_type = "SBS")
}
\arguments{
\item{x}{result from \link{sig_fit}.}

\item{result_dir}{a result directory.}

\item{mut_type}{one of 'SBS', 'DBS', 'ID' or 'CN'.}
}
\value{
Nothing.
}
\description{
Output Signature Fitting Results
}
