% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_sig}
\alias{output_sig}
\title{Output Signature Results}
\usage{
output_sig(sig, result_dir, mut_type = "SBS")
}
\arguments{
\item{sig}{a \code{Signature} object.}

\item{result_dir}{a result directory.}

\item{mut_type}{one of 'SBS', 'DBS', 'ID' or 'CN'.}
}
\value{
Nothing.
}
\description{
Output Signature Results
}
