% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{output_tally}
\alias{output_tally}
\title{Output Tally Result in Barplots}
\usage{
output_tally(x, result_dir, mut_type = "SBS")
}
\arguments{
\item{x}{a matrix with row representing components (motifs) and column
representing samples.}

\item{result_dir}{a result directory.}

\item{mut_type}{one of 'SBS', 'DBS', 'ID' or 'CN'.}
}
\value{
Nothing.
}
\description{
Output Tally Result in Barplots
}
