% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_copynumber_seqz.R
\name{read_copynumber_seqz}
\alias{read_copynumber_seqz}
\title{Read Absolute Copy Number Profile from Sequenza Result Directory}
\usage{
read_copynumber_seqz(target_dir, return_df = FALSE, ...)
}
\arguments{
\item{target_dir}{a directory path.}

\item{return_df}{if \code{TRUE}, return a \code{data.frame} directly, otherwise return a
\link{CopyNumber} object.}

\item{...}{other parameters passing to \code{\link[=read_copynumber]{read_copynumber()}}.}
}
\value{
a \code{data.frame} or a \code{CopyNumber} object.
}
\description{
Read Absolute Copy Number Profile from Sequenza Result Directory
}
