% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_fit_bootstrap.R
\name{report_bootstrap_p_value}
\alias{report_bootstrap_p_value}
\title{Report P Values from bootstrap Results}
\usage{
report_bootstrap_p_value(x, thresholds = c(0.01, 0.05, 0.1))
}
\arguments{
\item{x}{a (list of) result from \link{sig_fit_bootstrap}.}

\item{thresholds}{a vector of relative exposure threshold for calculating p values.}
}
\value{
a (list of) \code{matrix}
}
\description{
See examples in \link{sig_fit_bootstrap}.
}
