% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_sig_order.R
\name{reset_sig_names}
\alias{reset_sig_names}
\title{Reset Signature Names}
\usage{
reset_sig_names(Signature, names, order = TRUE)
}
\arguments{
\item{Signature}{a \code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract},
or just a raw signature matrix with row representing components (motifs) and column
representing signatures (column names must start with 'Sig').}

\item{names}{names to reset}

\item{order}{can be logical or a character vector.
If \code{TRUE}, signature names are sorted by \link{sort} function.}

\item{...}{other parameters passing to \link{sort}.}
}
\value{
a \code{Signature} object or a matrix.
}
\description{
Reset Signature Names
}
\examples{
print("yes")
}
