% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cn_components.R
\name{show_cn_components}
\alias{show_cn_components}
\title{Show Mixture Fit Model Components}
\usage{
show_cn_components(features, components, return_plotlist = FALSE,
  base_size = 12, ...)
}
\arguments{
\item{features}{a \code{list} generate from \link{derive} function.}

\item{components}{a \code{list} contain flexmix object of copy-number features, obtain this
from \link{derive} function.}

\item{return_plotlist}{if \code{TRUE}, return a list of ggplot objects but a combined plot.}

\item{base_size}{overall font size.}

\item{...}{other options pass to \code{\link[cowplot]{plot_grid}} function of \strong{cowplot} package.}
}
\value{
a \code{ggplot} object
}
\description{
Show mixture fit model components for copy number data.
}
\examples{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
show_cn_components(cn_prepare$features, cn_prepare$components)
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
