% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cn_features.R
\name{show_cn_features}
\alias{show_cn_features}
\title{Show Copy Number Feature Distributions}
\usage{
show_cn_features(
  features,
  method = "Macintyre",
  rm_outlier = FALSE,
  ylab = NULL,
  log_segsize = TRUE,
  return_plotlist = FALSE,
  base_size = 12,
  ...
)
}
\arguments{
\item{features}{a feature \code{list} generate from \link{sig_feed} function.}

\item{method}{method for feature classfication, can be one of "Macintyre" ("M") and
"Wang" ("W").}

\item{rm_outlier}{default is \code{FALSE}, if \code{TRUE}, remove outliers. Only
works when method is "Wang" ("W").}

\item{ylab}{lab of y axis.}

\item{log_segsize}{default is \code{TRUE}, show \code{log10} based segsize, only
works for input from "Macintyre" ("M") method.}

\item{return_plotlist}{if \code{TRUE}, return a list of ggplot objects but a combined plot.}

\item{base_size}{overall font size.}

\item{...}{other options pass to \code{\link[cowplot]{plot_grid}} function of \code{cowplot} package.}
}
\value{
a \code{ggplot} object
}
\description{
Show Copy Number Feature Distributions
}
\examples{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
show_cn_features(cn_prepare$features)
}
