% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cn_freq_circos.R
\name{show_cn_freq_circos}
\alias{show_cn_freq_circos}
\title{Show Copy Number Variation Frequency Profile with Circos}
\usage{
show_cn_freq_circos(
  data,
  groups = NULL,
  cutoff = 2L,
  title = c("AMP", "DEL"),
  chrs = paste0("chr", 1:22),
  genome_build = c("hg19", "hg38"),
  cols = NULL,
  ...
)
}
\arguments{
\item{data}{a \code{CopyNumber} object or a data.frame containing
at least 'chromosome', 'start', 'end', 'segVal', 'sample' these columns.}

\item{groups}{a named list or a column name for specifying groups.}

\item{cutoff}{copy number value cutoff for splitting data into AMP and DEL.
The values equal to cutoff are discarded. Default is \code{2}.}

\item{title}{length-2 titles for AMP and DEL.}

\item{chrs}{chromosomes start with 'chr'.}

\item{genome_build}{genome build version, used when \code{data} is a \code{data.frame}, should be 'hg19' or 'hg38'.}

\item{cols}{length-2 colors for AMP and DEL.}

\item{...}{other parameters passing to \link[circlize:circos.genomicLines]{circlize::circos.genomicLines}.}
}
\value{
Nothing.
}
\description{
Show Copy Number Variation Frequency Profile with Circos
}
\examples{
load(system.file("extdata", "toy_copynumber.RData",
  package = "sigminer", mustWork = TRUE
))

show_cn_freq_circos(cn)
ss <- unique(cn@data$sample)
show_cn_freq_circos(cn, groups = list(a = ss[1:5], b = ss[6:10]), cols = c("red", "green"))
}
