% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cosmic_signature.R
\name{show_cosmic}
\alias{show_cosmic}
\title{Show Signature Information in Web Browser}
\usage{
show_cosmic(x = "home")
}
\arguments{
\item{x}{a string indicating location
("home" for COSMIC signature home, "legacy" for COSMIC v2 signatures,
"SBS" for COSMIC v3 SBS signatures, "DBS" for COSMIC v3 DBS signatures,
"ID" for COSMIC v3 INDEL signatures) or signature index (e.g.
"SBS1", "DBS2", "ID3").}
}
\value{
Nothing.
}
\description{
Show Signature Information in Web Browser
}
\examples{
\dontrun{
show_cosmic()
show_cosmic("legacy")
show_cosmic("SBS")
show_cosmic("DBS")
show_cosmic("ID")
show_cosmic("SBS1")
show_cosmic("DBS2")
show_cosmic("ID3")
}
}
