% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_groups.R
\name{show_groups}
\alias{show_groups}
\title{Show Signature Contribution in Clusters}
\usage{
show_groups(grp_dt, ...)
}
\arguments{
\item{grp_dt}{a result \code{data.table} from \link{get_groups}.}

\item{...}{parameters passing to \code{\link[=legend]{legend()}}, e.g. \code{x = "topleft"}.}
}
\value{
nothing.
}
\description{
See example section in \code{\link[=sig_fit]{sig_fit()}} for an examples.
}
\seealso{
\link{get_groups}, \link{sig_fit}.
}
