% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sig_feature_corrplot.R
\name{show_sig_feature_corrplot}
\alias{show_sig_feature_corrplot}
\title{Draw Corrplot for Signature Exposures and Other Features}
\usage{
show_sig_feature_corrplot(
  tidy_cor,
  feature_list,
  sort_features = FALSE,
  return_plotlist = FALSE,
  xlab = "Signatures",
  ylab = "Features",
  gradient_colors = c("blue", "white", "red"),
  align = c("row", "column"),
  plot_ratio = "auto",
  breaks_count = c(0L, 200L, 400L, 600L, 800L, 1020L)
)
}
\arguments{
\item{tidy_cor}{data returned by \link{get_tidy_association}.}

\item{feature_list}{a character vector contains features want to be plotted.
If missing, all features will be used.}

\item{sort_features}{default is \code{FALSE}, use feature order obtained from the previous
step. If \code{TRUE}, sort features as \code{feature_list}.}

\item{return_plotlist}{if \code{TRUE}, return as a list of \code{ggplot} objects.}

\item{xlab}{label for x axis.}

\item{ylab}{label for y axis.}

\item{gradient_colors}{gradient colors used to plot.}

\item{align}{one of "row" and "column".}

\item{plot_ratio}{a length-2 numeric vector to set the height/width ratio.}

\item{breaks_count}{breaks for sample count.}
}
\value{
a \code{ggplot2} object
}
\description{
Draw Corrplot for Signature Exposures and Other Features
}
\seealso{
\link{get_tidy_association} and \link{get_sig_feature_association}
}
