% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_series.R
\name{sig_assign_samples}
\alias{sig_assign_samples}
\title{Return sample clustering from NMF run results}
\usage{
sig_assign_samples(nmfObj, type = "consensus", matchConseOrder = F)
}
\arguments{
\item{nmfObj}{a \code{NMF} result object which is an element return from \link{sig_extract}
or run results of \strong{NMF} package.}

\item{type}{cluster type, could be 'consensus' or 'sample'.}

\item{matchConseOrder}{if \code{TRUE}, the result will match order as shown in consensus map.}
}
\value{
a \code{data.frame}
}
\description{
One of key results from NMF decomposition is to cluster samples into different
groups. This function takes NMF result (a \code{NMF} object) as input
and return the membership in each cluster.
}
\details{
X = W x H

W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix.

More detail please see \code{\link[NMF:predict]{NMF::predict()}}.
}
\examples{
\donttest{
load(system.file("extdata", "example_cn_list.RData",
     package = "sigminer", mustWork = TRUE))
segTabs = data.table::rbindlist(tcga_segTabs, idcol = "sample")
cn = read_copynumber(segTabs,
                     seg_cols = c("chromosome", "start", "end", "segVal"),
                     genome_build = "hg19")
cn_prepare =  sig_prepare(cn)
res = sig_extract(cn_prepare$nmf_matrix, 3, mode = "copynumber", nrun = 5)
subtypes = sig_assign_samples(res$nmfObj)
}
}
\seealso{
Other signature analysis series function: \code{\link{sig_estimate}},
  \code{\link{sig_extract}},
  \code{\link{sig_get_activity}},
  \code{\link{sig_get_correlation}},
  \code{\link{sig_prepare}},
  \code{\link{sig_summarize_subtypes}}
}
\concept{signature analysis series function}
