% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_fit.R
\name{sig_fit}
\alias{sig_fit}
\title{Fit Signature Exposure with Linear Combination Decomposition}
\usage{
sig_fit(
  catalogue_matrix,
  sig,
  sig_index = NULL,
  sig_db = "legacy",
  show_index = TRUE,
  type = c("absolute", "relative"),
  return_class = c("matrix", "data.table"),
  rel_threshold = 0,
  mode = c("SBS", "copynumber")
)
}
\arguments{
\item{catalogue_matrix}{a numeric matrix \code{V} with row representing components and
columns representing samples, typically you can get \code{nmf_matrix} from \code{sig_tally()} and
transpose it by \code{t()}.}

\item{sig_index}{a vector for signature index.}

\item{sig_db}{can be 'legacy' or 'SBS'. Default 'legacy'.}

\item{show_index}{if \code{TRUE}, show valid index.}

\item{type}{'absolute' for signature exposure and 'relative' for signature relative exposure.}

\item{return_class}{string, 'matrix' or 'data.table'.}

\item{rel_threshold}{numeric vector, a relative exposure lower than this value will be set to 0.}

\item{mode}{signature type for plotting, now supports 'copynumber' or 'SBS'.}
}
\value{
The exposure result either in \code{matrix} or \code{data.table} format.
}
\description{
The function performs a signatures decomposition of a given mutational
catalogue \code{V} with known signatures \code{W} by solving the minimization problem
\verb{min(||W*H - V||)} with additional constraints of non-negativity on H where W and V are known.
}
\details{
This is a modification based on \code{LCD} function from YAPSA pakcage.
}
\examples{
W <- matrix(c(1, 2, 3, 4, 5, 6), ncol = 2)
colnames(W) <- c("sig1", "sig2")
W <- apply(W, 2, function(x) x / sum(x))

H <- matrix(c(2, 5, 3, 6, 1, 9, 1, 2), ncol = 4)
colnames(H) <- paste0("samp", 1:4)

V <- W \%*\% H
V

H_infer <- sig_fit(V, W)
H_infer
H

H_dt <- sig_fit(V, W, return_class = "data.table")
H_dt
}
