% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_series.R
\name{sig_group_enrichment}
\alias{sig_group_enrichment}
\title{Performs alteration enrichment analysis for a given group}
\usage{
sig_group_enrichment(data, ft_cols, group_col, positive_flag = TRUE,
  id_col = "Tumor_Sample_Barcode")
}
\arguments{
\item{data}{a \code{data.frame} contains group column, feature columns.}

\item{ft_cols}{a character vector for storing column name of features}

\item{group_col}{column name for group}

\item{positive_flag}{the flag indicating the feature is alterated in sample, e.g., \code{TRUE} for mutant and
\code{FALSE} for WT}

\item{id_col}{column name for sample id}
}
\value{
a \code{list}
}
\description{
Performs paiwise and groupwise fisher exact tests to find differentially enriched features
for every factor within a group. This is a modified version of \code{clinicalEnrichment()}
function from \strong{maftools} package. Of note, this function only works when feature has
2 factors. The result can be passed to \code{plotEnrichmentResults()} function of \strong{maftools} package.
}
\seealso{
\code{\link[maftools:plotEnrichmentResults]{maftools::plotEnrichmentResults()}}

Other signature analysis series function: \code{\link{sig_assign_samples}},
  \code{\link{sig_estimate}}, \code{\link{sig_extract}},
  \code{\link{sig_get_activity}},
  \code{\link{sig_get_correlation}},
  \code{\link{sig_get_similarity}},
  \code{\link{sig_prepare}},
  \code{\link{sig_summarize_subtypes}}
}
\concept{signature analysis series function}
