% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_correlation.R
\name{sig_summarize_correlation}
\alias{sig_summarize_correlation}
\title{Calculate association between signature activity and other features}
\usage{
sig_summarize_correlation(data, cols_to_sigs, cols_to_summary,
  type = "ca", method_co = c("spearman", "pearson", "kendall"),
  method_ca = stats::wilcox.test, min_n = 0.01, verbose = FALSE, ...)
}
\arguments{
\item{data}{a \code{data.frame} contains signature activity and other features}

\item{cols_to_sigs}{colnames for signature activity}

\item{cols_to_summary}{colnames for other features}

\item{type}{a character vector contains 'ca' for categorical variable and 'co' for continuous variable,
it must have same length as \code{cols_to_summary}.}

\item{method_co}{method for continuous variable, default is "spearman", could also be "pearson" and "kendall".}

\item{method_ca}{method for categorical variable, default is "wilcox.test"}

\item{min_n}{a minimal fraction (e.g. 0.01) or a integer number (e.g. 10) for filtering some variables with few positive events.
Default is 0.01.}

\item{verbose}{if \code{TRUE}, print extra message.}

\item{...}{other arguments passing to}
}
\value{
a \code{list}
}
\description{
Calculate association between signature activity and other features
}
