% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_correlation.R
\name{sig_tidy_correlation}
\alias{sig_tidy_correlation}
\title{Tidy signature correlation results}
\usage{
sig_tidy_correlation(cor_res)
}
\arguments{
\item{cor_res}{data returned by \code{\link[=sig_summarize_correlation]{sig_summarize_correlation()}}}
}
\value{
a \code{data.frame}
}
\description{
Tidy signature correlation results
}
