% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminer.R
\docType{package}
\name{sigminer-package}
\alias{sigminer}
\alias{sigminer-package}
\title{sigminer: Extract, Analyze and Visualize Signatures for Genomic Variations}
\description{
\itemize{
\item Author: \href{https://shixiangwang.github.io/home/}{Shixiang Wang} (\href{mailto:w_shixiang@163.com}{w_shixiang@163.com})
\item Please go to \url{https://shixiangwang.github.io/sigminer-doc/} for full vignette.
\item Please go to \url{https://shixiangwang.github.io/sigminer/reference/index.html}
for organized documentation of functions and datasets.
\item Result visualization for \link{MAF} is provide by \strong{maftools} package,
please read its \href{https://bioconductor.org/packages/release/bioc/vignettes/maftools/inst/doc/maftools.html}{vignette}.
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShixiangWang/sigminer}
  \item \url{https://shixiangwang.github.io/sigminer/}
  \item \url{https://shixiangwang.github.io/sigminer-book/}
  \item Report bugs at \url{https://github.com/ShixiangWang/sigminer/issues}
}

}
\author{
\strong{Maintainer}: Shixiang Wang \email{w_shixiang@163.com} (\href{https://orcid.org/0000-0001-9855-7357}{ORCID})

Authors:
\itemize{
  \item Ziyu Tao \email{taozy@shanghaitech.edu.cn} (\href{https://orcid.org/0000-0003-3272-1227}{ORCID})
  \item Huimin Li \email{lihm@shanghaitech.edu.cn} (\href{https://orcid.org/0000-0003-1683-9057}{ORCID})
  \item Tao Wu \email{wutao2@shanghaitech.edu.cn} (\href{https://orcid.org/0000-0002-8999-9628}{ORCID})
  \item Xue-Song Liu (\href{https://orcid.org/0000-0002-7736-0077}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Anand Mayakonda [contributor]
}

}
\keyword{internal}
