% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminer.R
\docType{package}
\name{sigminer}
\alias{sigminer}
\alias{sigminer-package}
\title{sigminer: Capture Genomic Variation Signatures using Non-Negative Matrix Factorization}
\description{
Genomic variations including somatic mutations, copy number variations, etc. are the major
force for cancer initialization and development. Due to the specificity of molecular lesions
caused by genomic variations, we can generate characteristic variation spectra, called 'signature'.
This package helps users to extract, analyze and visualize signatures from genomic variations,
thus providing new insight into cancer study.
}
\section{read data}{

Read data into R objects:
\itemize{
\item \code{\link[=read_copynumber]{read_copynumber()}} - read \code{data.frame}/files into \link{CopyNumber} object.
\item \code{\link[=read_maf]{read_maf()}} - read \code{data.frame}/file into \link{MAF} object. This is powered by
\code{\link[maftools]{read.maf}} function in \strong{maftools} package.
}
}

\section{sig_ functions}{

Signature analysis for genomic variations:
}

\section{draw_ functions}{

Result visualization for copy number data:

Result visualization for signature analysis:

Result visualization for \link{MAF} is provide by \strong{maftools} package,
please read its \href{https://bioconductor.org/packages/release/bioc/vignettes/maftools/inst/doc/maftools.html}{vignette}.
}

