% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_catalogs}
\alias{simulated_catalogs}
\title{A List of Simulated SBS-96 Catalog Matrix}
\format{
A list of matrix
}
\source{
Generate from code under data_raw/
}
\description{
Data from \url{https://doi.org/10.1038/s43018-020-0027-5}.
5 simulated catalog datasets are used by the paper but only 4 are available.
}
\examples{
data(simulated_catalogs)
}
