% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{subset.CopyNumber}
\alias{subset.CopyNumber}
\title{Subsetting CopyNumber object}
\usage{
\method{subset}{CopyNumber}(x, subset = TRUE, ...)
}
\arguments{
\item{x}{a \link{CopyNumber} object to be subsetted.}

\item{subset}{logical expression indicating rows to keep.}

\item{...}{further arguments to be passed to or from other methods.
Useless here.}
}
\value{
a \link{CopyNumber} object
}
\description{
Subset \code{data} slot of \link{CopyNumber} object, un-selected rows will move to
dropoff.segs slot, annotation slot will update in the same way.
}
\author{
Shixiang Wang
}
