% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript.hg38}
\alias{transcript.hg38}
\title{Merged Transcript Location at Genome Build hg38}
\format{
A \code{data.table}
}
\source{
from GENCODE release v33.
}
\description{
Merged Transcript Location at Genome Build hg38
}
\examples{
data(transcript.hg38)
}
