% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript.mm10}
\alias{transcript.mm10}
\title{Merged Transcript Location at Genome Build mm10}
\format{
A \code{data.table}
}
\source{
from GENCODE release M25.
}
\description{
Merged Transcript Location at Genome Build mm10
}
\examples{
data(transcript.mm10)
}
