% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript.mm9}
\alias{transcript.mm9}
\title{Merged Transcript Location at Genome Build mm9}
\format{
A \code{data.table}
}
\source{
from UCSC \url{http://hgdownload.cse.ucsc.edu/goldenPath/mm9/database/transcriptome.txt.gz}
}
\description{
Merged Transcript Location at Genome Build mm9
}
\examples{
data(transcript.mm9)
}
