% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript.T2T}
\alias{transcript.T2T}
\title{Merged Transcript Location at Genome Build T2T}
\format{
A \code{data.table}
}
\source{
from T2T study.
}
\description{
Merged Transcript Location at Genome Build T2T
}
\examples{
data(transcript.T2T)
}
