% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_color_style}
\alias{use_color_style}
\title{Set Color Style for Plotting}
\usage{
use_color_style(style)
}
\arguments{
\item{style}{one of 'default' and 'cosmic'.}
}
\value{
color values.
}
\description{
Set Color Style for Plotting
}
\examples{
use_color_style("default")
use_color_style("cosmic")
}
