% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.hg19}
\alias{centromeres.hg19}
\title{location of centromeres at genome build hg19}
\format{A data.frame}
\source{
Generate from UCSC gold path
}
\description{
location of centromeres at genome build hg19
}
\examples{
data(centromeres.hg19)
}
