% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centromeres.hg38}
\alias{centromeres.hg38}
\title{location of centromeres at genome build hg38}
\format{A data.frame}
\source{
Generate from Genome Reference Consortium
}
\description{
location of centromeres at genome build hg38
}
\examples{
data(centromeres.hg38)
}
