% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromsize.hg38}
\alias{chromsize.hg38}
\title{chromosome size of genome build hg38}
\format{A data.frame}
\source{
Generate from UCSC gold path
}
\description{
chromosome size of genome build hg38
}
\examples{
data(chromsize.hg38)
}
