% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CN.features}
\alias{CN.features}
\title{Classfication Table of Copy Number Features Devised by Wang et al.}
\format{
A \code{data.table} with "sigminer.features" class name
}
\source{
Generate from code under data_raw/
}
\description{
Classfication Table of Copy Number Features Devised by Wang et al.
}
\examples{
data(CN.features)
}
