% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{draw_cn_features}
\alias{draw_cn_features}
\title{Plot copy number feature distribution}
\usage{
draw_cn_features(features, ylab = "", ...)
}
\arguments{
\item{features}{a \code{list} generate from \link{get_features} or \link{sig_prepare} function.}

\item{ylab}{lab of y axis.}

\item{...}{other options pass to \code{\link[cowplot]{plot_grid}} function of \code{cowplot} package.}
}
\value{
a ggplot object
}
\description{
Plot copy number feature distribution
}
\examples{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
draw_cn_features(cn_prepare$features)
}
\seealso{
Other copy number plot: \code{\link{draw_cn_components}},
  \code{\link{draw_cn_distribution}}
}
\concept{copy number plot}
