% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_series.R
\name{draw_sig_profile}
\alias{draw_sig_profile}
\title{Plot signature profile}
\usage{
draw_sig_profile(nmfObj, mode = c("copynumber", "mutation"),
  params = NULL, params_label_size = 2, y_expand = 1, digits = 1,
  y_scale = c("relative", "absolute"), font_scale = 1,
  sig_names = NULL, sig_orders = NULL)
}
\arguments{
\item{nmfObj}{a \code{NMF} result object which is an element return from \link{sig_extract}
or run results of \strong{NMF} package.}

\item{mode}{variation type to decompose, currently support "copynumber" or "mutation".}

\item{params}{params data of components.}

\item{params_label_size}{font size for params label.}

\item{y_expand}{y expand height for plotting params of copy number signatures.}

\item{digits}{digits for plotting params of copy number signatures.}

\item{y_scale}{one of 'relative' or 'absolute', if choose 'relative',
signature columns will be scaled to sum as 1.}

\item{font_scale}{a number used to set font scale.}

\item{sig_names}{set name of signatures, can be a character vector.
Default is \code{NULL}, prefix 'Signature_' plus number is used.}

\item{sig_orders}{set order of signatures, can be a character vector.
Default is \code{NULL}, the signatures are ordered by alphabetical order.}
}
\value{
a \code{ggplot} object
}
\description{
Currently support copy number signatures and mutation signatures.
}
\examples{
# Load copy number signature
load(system.file("extdata", "toy_copynumber_signature.RData",
  package = "sigminer", mustWork = TRUE
))
draw_sig_profile(res$nmfObj)
}
\seealso{
Other signature plot: \code{\link{draw_sig_activity}},
  \code{\link{draw_sig_corrplot}},
  \code{\link{draw_subtypes_comparison}}
}
\author{
Shixiang Wang
}
\concept{signature plot}
