% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_cnlist}
\alias{get_cnlist}
\title{Extract copy number profile as list from CopyNumber object}
\usage{
get_cnlist(CopyNumber)
}
\arguments{
\item{CopyNumber}{a \link{CopyNumber} object.}
}
\value{
a \code{list}
}
\description{
Extract copy number profile as list from CopyNumber object
}
\examples{
extdata_dir <- system.file("extdata", package = "sigminer", mustWork = TRUE)
cp <- read_copynumber(extdata_dir, pattern = "txt", genome_build = "hg19")
cn_list <- get_cnlist(cp)
}
\seealso{
Other internal calculation function series: \code{\link{get_ArmLocation}},
  \code{\link{get_LengthFraction}},
  \code{\link{get_cnsummary_sample}},
  \code{\link{get_components}}, \code{\link{get_features}},
  \code{\link{get_matrix}}
}
\author{
Shixiang Wang
}
\concept{internal calculation function series}
