% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_series.R
\name{get_LengthFraction}
\alias{get_LengthFraction}
\title{Calculate length fraction profile of copy number}
\usage{
get_LengthFraction(CN_data, genome_build = c("hg19", "hg38"),
  seg_cols = c("Chromosome", "Start.bp", "End.bp", "modal_cn"),
  samp_col = "sample", check_chrom = FALSE)
}
\arguments{
\item{CN_data}{a \code{list} contains multiple \code{data.frame}s (recommended),
each \code{data.frame} stores copy-number profile for one sample with 'chromosome', 'start', 'end' and
'segVal' these four necessary columns. Of note, 'segVal' column shoule be absolute copy number values.}

\item{genome_build}{genome build version, should be 'hg19' or 'hg38'.}

\item{seg_cols}{four characters used to specify chromosome, start position,
end position and copy number value in \code{input}, respectively.
Default use names from ABSOLUTE calling result.}

\item{samp_col}{a character used to specify the sample column name. If \code{input}
is a directory and cannot find \code{samp_col}, sample names will use file names
(set this parameter to \code{NULL} is recommended in this case).}

\item{check_chrom}{if \code{TRUE}, check chromosome name, name like '23' or 'chr23'
will be discarded.}
}
\value{
a data frame
}
\description{
Calculate length fraction profile of copy number
}
\seealso{
Other internal calculation function series: \code{\link{get_ArmLocation}},
  \code{\link{get_components}}, \code{\link{get_features}},
  \code{\link{get_matrix}}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
\concept{internal calculation function series}
