% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sig_similarity.R
\name{get_sig_db}
\alias{get_sig_db}
\title{Obtain Reference Signatures}
\usage{
get_sig_db(sig_db = "legacy")
}
\arguments{
\item{sig_db}{can be 'legacy' (for COSMIC v2 'SBS'),
'SBS', 'DBS', 'ID' and 'TSB' (for SBS transcriptional strand bias signatures).
Default 'legacy'.}
}
\value{
a \code{list}.
}
\description{
The signatures and their aetiologies mainly obtained from COSMIC database and cleaned before saving into
\strong{sigminer} package.
}
\examples{
s1 <- get_sig_db()
s2 <- get_sig_db("DBS")
s1
s2
}
\seealso{
\link{get_sig_similarity}, \link{sig_fit} and \link{show_cosmic_sig_profile}.
}
