% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_cn_components.R
\name{show_cn_components}
\alias{show_cn_components}
\title{Show Mixture Fit Model Components}
\usage{
show_cn_components(
  parameters,
  show_weights = TRUE,
  log_segsize = TRUE,
  auto_transform = TRUE,
  return_plotlist = FALSE,
  base_size = 12,
  ...
)
}
\arguments{
\item{parameters}{a \code{data.frame} contain parameter components, obtain this
from \link{sig_derive} function.}

\item{show_weights}{default is \code{TRUE}, show weights for each component.}

\item{log_segsize}{default is \code{TRUE}, show \code{log10} based segsize.}

\item{auto_transform}{default is \code{TRUE}, it will auto increase the \code{SD} for components
for showing them better in the plot.}

\item{return_plotlist}{if \code{TRUE}, return a list of ggplot objects but a combined plot.}

\item{base_size}{overall font size.}

\item{...}{other options pass to \code{\link[cowplot]{plot_grid}} function of \strong{cowplot} package.}
}
\value{
a \code{ggplot} object
}
\description{
Show mixture fit model components for copy number data.
}
\examples{
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
show_cn_components(cn_prepare$parameters)
show_cn_components(cn_prepare$parameters, show_weights = FALSE)
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
