% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_group_distribution.R
\name{show_group_distribution}
\alias{show_group_distribution}
\title{Show Groupped Variable Distribution}
\usage{
show_group_distribution(
  data,
  gvar,
  dvar,
  g_label = NULL,
  xlab = NULL,
  ylab = NULL,
  background_color = "#DBD7D2"
)
}
\arguments{
\item{data}{a \code{data.frame}.}

\item{gvar}{a group variable name/index.}

\item{dvar}{a distribution variable name/index.}

\item{g_label}{a named vector to set facet labels, default is \code{NULL}.}

\item{xlab}{title for x axis.}

\item{ylab}{title for y axis.}

\item{background_color}{background color for plot panel.}
}
\value{
a \code{ggplot} object.
}
\description{
This is a general function, it can be used in any proper analysis.
}
\examples{
set.sed(1234)
data <- data.frame(
  y = rnorm(120),
  g = c(rep("A", 50), rep("B", 40), rep("C", 30))
)
p <- show_group_distribution(data, gvar = 2, dvar = 1, background_color = "grey")
p
}
