% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sig_exposure.R
\name{show_sig_exposure}
\alias{show_sig_exposure}
\title{Plot Signature Exposure}
\usage{
show_sig_exposure(
  Signature,
  cutoff = NULL,
  base_size = 12,
  font_scale = 1,
  rm_space = FALSE,
  hide_samps = TRUE
)
}
\arguments{
\item{Signature}{a \code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract},
or just a raw exposure matrix with column representing samples (patients) and row
representing signatures (row names must start with 'Sig').}

\item{cutoff}{a cutoff value to remove hyper-mutated samples.}

\item{base_size}{overall font size.}

\item{font_scale}{a number used to set font scale.}

\item{rm_space}{default is \code{FALSE}. If \code{TRUE}, it will remove border color
and expand the bar width to 1. This is useful when the sample size is big.}

\item{hide_samps}{if \code{TRUE}, hide sample names.}
}
\value{
a \code{ggplot} object
}
\description{
Currently support copy number signatures and mutational signatures.
}
\examples{
# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Show signature exposure
show_sig_exposure(sig2)

# Load copy number signature
load(system.file("extdata", "toy_copynumber_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Show signature exposure
show_sig_exposure(sig)
}
\author{
Shixiang Wang
}
