% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sig_feature_corrplot.R
\name{show_sig_feature_corrplot}
\alias{show_sig_feature_corrplot}
\title{Draw Corrplot for Signature Exposures and Other Features}
\usage{
show_sig_feature_corrplot(
  tidy_cor,
  features_list,
  breaks_count = c(0L, 200L, 400L, 600L, 800L, 1020L),
  breaks_score = c(-Inf, -16, -10, -5, -1.3, 1.3, 5, 10, 16, Inf)
)
}
\arguments{
\item{tidy_cor}{data returned by \link{get_tidy_association}.}

\item{features_list}{a character vector contains features want to be plotted.}

\item{breaks_count}{breaks for sample count.}

\item{breaks_score}{breaks for score.}
}
\value{
a \code{ggplot2} object
}
\description{
Draw Corrplot for Signature Exposures and Other Features
}
\seealso{
\link{get_tidy_association} and \link{get_sig_feature_association}
}
