% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sig_profile.R
\name{show_sig_profile}
\alias{show_sig_profile}
\title{Show Signature Profile}
\usage{
show_sig_profile(
  Signature,
  mode = c("copynumber", "mutation"),
  method = "Macintyre",
  normalize = c("row", "column", "raw"),
  feature_setting = sigminer::CN.features,
  x_label_angle = 60,
  params = NULL,
  show_cv = FALSE,
  params_label_size = 3,
  params_label_angle = 60,
  y_expand = 1,
  digits = 2,
  base_size = 12,
  font_scale = 1,
  sig_names = NULL,
  sig_orders = NULL,
  check_sig_names = TRUE
)
}
\arguments{
\item{Signature}{a \code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract},
or just a raw signature matrix with row representing components (motifs) and column
representing signatures (column names must start with 'Sig').}

\item{mode}{signature type for plotting, now supports 'copynumber' or 'mutation'.}

\item{method}{method for copy number feature classfication, can be one of "Macintyre" ("M") and
"Wang" ("W").}

\item{normalize}{one of 'row', 'column' and 'raw', for row normalization (signature),
column normalization (component) and raw data, respectively.}

\item{feature_setting}{a \code{data.frame} used for classification.
Only used when method is "Wang" ("W").
Default is \link{CN.features}. Users can also set custom input with "feature",
"min" and "max" columns avaiable. Valid features can be printed by
\code{unique(CN.features$feature)}.}

\item{x_label_angle}{font angle for x label.}

\item{params}{params \code{data.frame} of components, obtained from \link{sig_derive}.}

\item{show_cv}{default is \code{FALSE}, if \code{TRUE}, show coefficient of variation when
\code{params} is not \code{NULL}.}

\item{params_label_size}{font size for params label.}

\item{params_label_angle}{font angle for params label.}

\item{y_expand}{y expand height for plotting params of copy number signatures.}

\item{digits}{digits for plotting params of copy number signatures.}

\item{base_size}{overall font size.}

\item{font_scale}{a number used to set font scale.}

\item{sig_names}{set name of signatures, can be a character vector.
Default is \code{NULL}, prefix 'Sig_' plus number is used.}

\item{sig_orders}{set order of signatures, can be a character vector.
Default is \code{NULL}, the signatures are ordered by alphabetical order.}

\item{check_sig_names}{if \code{TRUE}, check signature names when input is
a matrix, i.e., all signatures (colnames) must start with 'Sig'.}
}
\value{
a \code{ggplot} object
}
\description{
Who don't like to show a barplot for signature profile? This is for it.
}
\examples{
# Load mutational signature
load(system.file("extdata", "toy_mutational_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Show signature profile
show_sig_profile(sig2, mode = "mutation")


# Load copy number signature
load(system.file("extdata", "toy_copynumber_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Show signature profile
show_sig_profile(sig)

# Add params label
# =================
# Load copy number prepare object
load(system.file("extdata", "toy_copynumber_prepare.RData",
  package = "sigminer", mustWork = TRUE
))
params <- get_tidy_parameter(cn_prepare$components)
show_sig_profile(sig, params = params, y_expand = 2)
}
\author{
Shixiang Wang
}
