% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sig_profile.R
\name{show_sig_profile}
\alias{show_sig_profile}
\title{Show Signature Profile}
\usage{
show_sig_profile(Signature, mode = c("copynumber", "mutation"),
  params = NULL, params_label_size = 3, y_expand = 1, digits = 1,
  base_size = 12, y_scale = c("relative", "absolute"),
  font_scale = 1, sig_names = NULL, sig_orders = NULL)
}
\arguments{
\item{mode}{either "ld" for distribution by CN length or "cd" for distribution by chromosome.}

\item{params}{params data of components.}

\item{params_label_size}{font size for params label.}

\item{y_expand}{y expand height for plotting params of copy number signatures.}

\item{digits}{digits for plotting params of copy number signatures.}

\item{base_size}{overall font size.}

\item{y_scale}{one of 'relative' or 'absolute', if choose 'relative',
signature columns will be scaled to sum as 1.}

\item{font_scale}{a number used to set font scale.}

\item{sig_names}{set name of signatures, can be a character vector.
Default is \code{NULL}, prefix 'Signature_' plus number is used.}

\item{sig_orders}{set order of signatures, can be a character vector.
Default is \code{NULL}, the signatures are ordered by alphabetical order.}

\item{a}{\code{Signature} object obtained either from \link{sig_extract} or \link{sig_auto_extract}.}
}
\value{
a \code{ggplot} object
}
\description{
Who don't like to show a barplot for signature profile? This is for it.
}
\examples{
\donttest{
# Load copy number signature
#// Remove this data and create a new one
load(system.file("extdata", "toy_copynumber_signature.RData",
  package = "sigminer", mustWork = TRUE
))
show_sig_profile(res$nmfObj)
}
}
\author{
Shixiang Wang
}
