% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_series.R
\name{sig_get_activity}
\alias{sig_get_activity}
\title{Get signature activity}
\usage{
sig_get_activity(nmfObj)
}
\arguments{
\item{nmfObj}{a \code{NMF} result object which is an element return from \link{sig_extract}
or run results of \strong{NMF} package.}
}
\value{
a \code{list}
}
\description{
Get signature activity
}
\examples{
# Load copy number signature
load(system.file("extdata", "toy_copynumber_signature.RData",
  package = "sigminer", mustWork = TRUE
))
# Get activity of signatures
sig_activity <- sig_get_activity(res$nmfObj)
}
\seealso{
Other signature analysis series function: \code{\link{sig_assign_samples}},
  \code{\link{sig_estimate}}, \code{\link{sig_extract}},
  \code{\link{sig_get_correlation}},
  \code{\link{sig_get_similarity}},
  \code{\link{sig_prepare}},
  \code{\link{sig_summarize_subtypes}}
}
\concept{signature analysis series function}
