% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminer.R
\docType{package}
\name{sigminer}
\alias{sigminer}
\alias{sigminer-package}
\title{sigminer: Capture Genomic Variation Signatures using Non-Negative Matrix Factorization}
\description{
sigminer provides functions for identification of
copy number signatures (Geoffrey et al. (2018) \url{doi:10.1038/s41588-018-0179-8}) and
mutation signatures (Alexandrov et al. (2018) \url{doi:10.1038/nature12477})
by non-negative matrix factorization, signature analysis and visualization.
It can be used to capture signatures of genomic variation, compare genotype or
phenotype features of different signatures and thus uncover the relationship
between the mechanism of genomic variation and phenotypes in cancer.
}
\section{read_ functions}{

read_ functions read data into R objects.
}

