% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigminer.R
\docType{package}
\name{sigminer}
\alias{sigminer}
\title{sigminer: Extract, Analyze and Visualize Signatures for Genomic Variations}
\description{
Please go to \url{https://shixiangwang.github.io/sigminer-doc/} for full vignette.
}
\details{
Result visualization for \link{MAF} is provide by \strong{maftools} package,
please read its \href{https://bioconductor.org/packages/release/bioc/vignettes/maftools/inst/doc/maftools.html}{vignette}.
}
